+++ src/main/java/life/echo/community/controller/QusetionController.java
@@ -15,7 +15,7 @@ public class QusetionController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/question/{id}")
-    public String Question(@PathVariable(name="id")Integer id, Model model) {
+    public String Question(@PathVariable(name="id")Long id, Model model) {
         QuestionDTO questionDTO = questionService.getById(id);
         questionService.incView(id);
         model.addAttribute("question",questionDTO);
